# backend/api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    JournalEntry:
      properties:
        ai_insights:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Ai Insights
        content:
          minLength: 1
          title: Content
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          title: Id
          type: string
        mood_indicators:
          items:
            type: string
          title: Mood Indicators
          type: array
        mood_score:
          anyOf:
          - maximum: 10.0
            minimum: 1.0
            type: integer
          - type: 'null'
          title: Mood Score
        sentiment_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Sentiment Score
        updated_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updated At
        user_id:
          title: User Id
          type: string
      required:
      - content
      - user_id
      title: JournalEntry
      type: object
    JournalEntryCreate:
      properties:
        content:
          minLength: 1
          title: Content
          type: string
        mood_indicators:
          items:
            type: string
          title: Mood Indicators
          type: array
        mood_score:
          anyOf:
          - maximum: 10.0
            minimum: 1.0
            type: integer
          - type: 'null'
          title: Mood Score
      required:
      - content
      title: JournalEntryCreate
      type: object
...
    User:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        email:
          format: email
          title: Email
          type: string
        id:
          title: Id
          type: string
        preferences:
          additionalProperties: true
          title: Preferences
          type: object
        profile:
          additionalProperties: true
          title: Profile
          type: object
        subscription_tier:
          default: free
          title: Subscription Tier
          type: string
      required:
      - email
      - id
      - created_at
      title: User
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: API for mental health tracking and mindfulness exercises
  title: Mental Health Companion API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/auth/user/user:
    get:
      description: Retrieve the current user's profile
      operationId: get_current_user_profile_api_auth_user_user_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful Response
        '400': &id001
          content:
            application/json:
              schema:
                properties:
                  detail:
                    type: string
                type: object
          description: Bad Request
        '401': &id002
          content:
            application/json:
              schema:
                properties:
                  detail:
                    type: string
                type: object
          description: Unauthorized
        '404': &id004
          content:
            application/json:
              schema:
                properties:
                  detail:
                    type: string
                type: object
          description: Not Found
        '500': &id003
          content:
            application/json:
              schema:
                properties:
                  detail:
                    type: string
                type: object
          description: Internal Server Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Current User Profile
      tags:
      - Authentication
...
  /api/journal/:
    get:
      description: Get all journal entries for the current user
      operationId: get_journal_entries_api_journal__get
      parameters:
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/JournalEntry'
                title: Response Get Journal Entries Api Journal  Get
                type: array
          description: Successful Response
        '400': *id001
        '401': *id002
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500': *id003
      security:
      - OAuth2PasswordBearer: []
      summary: Get Journal Entries
      tags:
      - Journal
    post:
      description: Create a new journal entry
      operationId: create_journal_entry_api_journal__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalEntryCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
          description: Successful Response
        '400': *id001
        '401': *id002
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500': *id003
      security:
      - OAuth2PasswordBearer: []
      summary: Create Journal Entry
      tags:
      - Journal
...